Move c-errors related code to api_lib module
