+++ libvcx/src/api_lib/utils_c/mod.rs
@@ -7,3 +7,4 @@ pub mod runtime;
 pub mod timeout;
 pub mod return_types_u32;
 pub mod fsm_states;
+pub mod error;
